+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next release
+* Strong mode breaking change: can now infer generic type arguments from the constructor invocation arguments (#25220).
+
 ## 0.27.4-alpha.16
 * (Internal) Corresponds with the analyzer/server in the `1.18.0-dev.4.0` SDK.